Move endpopints-health-check.ts
