+++ src/frontend/packages/core/src/core/endpoints-health-checks.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 
-import { entityCatalog } from '../store/src/entity-catalog/entity-catalog';
-import { EndpointHealthCheck } from '../store/src/entity-catalog/entity-catalog.types';
-import { EndpointModel } from '../store/src/types/endpoint.types';
+import { entityCatalog } from '../../../store/src/entity-catalog/entity-catalog';
+import { EndpointHealthCheck } from '../../../store/src/entity-catalog/entity-catalog.types';
+import { EndpointModel } from '../../../store/src/types/endpoint.types';
 
 @Injectable({
   providedIn: 'root'
